Refactor project structure: remove unused components (Backend, Creative, Frontend, Workflow), update button styles and classes, enhance Card component with Font Awesome icons, and improve layout for various sections.
