+++ src/Components/Buttons/GithubButton.jsx
@@ -2,7 +2,12 @@ import "./GithubButton.css";
 
 export const GithubButton = ({ url }) => {
   return (
-    <a href={url} target="_blank" rel="noopener noreferrer" className="button">
+    <a
+      href={url}
+      target="_blank"
+      rel="noopener noreferrer"
+      className="github links"
+    >
       Github
     </a>
   );
