+++ src/Data/projects.json
@@ -3,7 +3,7 @@
     {
       "index": 0,
       "name": "Business site",
-      "icon": "faFilm",
+      "icon": "film",
       "tags": ["HTML5", "CSS3", "JavaScript"],
       "description": "My first-ever website, built during my Technigo bootcamp using HTML5, CSS3, and basic JavaScript. Created independently over three weeks with a focus on responsive design and clean code.",
       "netlify": "https://midnight-matinee.netlify.app/",
@@ -11,8 +11,8 @@
     },
     {
       "index": 1,
-      "name": "Hack Week: Weather App",
-      "icon": "faCloudSunRain",
+      "name": "Team Weather App",
+      "icon": "cloud-sun-rain",
       "tags": ["HTML5", "CSS3", "JavaScript", "TypeScript", "APIs", "github"],
       "description": "Our first collaborative team project, created during Hack Week, using a weather API along with TypeScript and Geolocation. I was primarily responsible for implementing the CSS and reviewing/testing the TypeScript to ensure a smooth and responsive user experience",
       "netlify": "https://weatherapp-teamproject.netlify.app/",
@@ -21,7 +21,7 @@
     {
       "index": 2,
       "name": "Recipe Library API",
-      "icon": "faFoodApple",
+      "icon": "utensils",
       "tags": ["HTML5", "CSS3", "JavaScript", "APIs", "github"],
       "description": "A responsive web application I designed and developed, allowing users to explore recipes based on cuisine, diet, or keyword, powered by the Spoonacular API. Using a Figma prototype as the foundation, I made personalized UI enhancements to optimize the mobile experience. Built with HTML, CSS, and JavaScript, the app features filtering, sorting, and a random recipe generator, offering an engaging and seamless user experience across devices.",
       "netlify": "https://keepcooking.netlify.app/",
@@ -29,8 +29,8 @@
     },
     {
       "index": 3,
-      "name": "Acecessibility App",
-      "icon": "faAccessibility",
+      "name": "Team Acecessibility App",
+      "icon": "universal-access",
       "tags": [
         "HTML5",
         "CSS3",
