+++ src/Sections/MainBody/Projects/Projects.css
@@ -1,4 +1,4 @@
-#projects {
+.projects {
   display: flex;
   flex-direction: row;
   justify-content: center;
