+++ src/Sections/Skills.jsx
@@ -1,7 +1,9 @@
 export const Skills = () => {
   return (
     <section id="skills">
-      <h2>Skills</h2>
+      <div className="h2-title">
+        <h2>Skills</h2>
+      </div>
     </section>
   );
 };
