+++ src/Sections/Toolbox.jsx
@@ -1,13 +1,15 @@
-import { Backend } from "../Components/Backend";
-import { Frontend } from "../Components/Frontend";
-import { Workflow } from "../Components/Workflow";
-import { Creative } from "../Components/Creative";
+import { Backend } from "../Components/TechLists/Backend";
+import { Frontend } from "../Components/TechLists/Frontend";
+import { Workflow } from "../Components/TechLists/Workflow";
+import { Creative } from "../Components/TechLists/Creative";
 import "./Toolbox.css";
 
 export const Toolbox = () => {
   return (
     <section id="toolbox">
-      <h2> My Toolbox</h2>
+      <div className="h2-title">
+        <h2> My Toolbox</h2>
+      </div>
       <div className="toolbox-list">
         <Frontend />
         <Backend />
