+++ src/index.css
@@ -33,7 +33,7 @@ body {
 h1 {
   color: var(--font-color-light);
   font-family: var(--font-Header);
-  font-size: 6rem;
+  font-size: 6.5rem;
   font-weight: 700;
   text-shadow: var(--shadow);
   margin: 0.5rem;
@@ -49,14 +49,14 @@ h4 {
 }
 
 h2 {
-  font-size: 2rem;
+  font-size: 3rem;
   font-weight: 800;
   text-shadow: var(--shadow);
   margin: 0.5rem;
 }
 
 h3 {
-  font-size: 1.25rem;
+  font-size: 2rem;
   font-weight: 600;
 }
 h4 {
@@ -82,16 +82,6 @@ a {
   text-decoration: none;
 }
 
-svg {
-  color: var(--color-accent);
-}
-
-svg:hover {
-  color: var(--color-neutral);
-  transform: scale(1.05);
-  transition: all 0.3s ease-in-out;
-}
-
 section {
   display: flex;
   flex-direction: column;
@@ -99,6 +89,13 @@ section {
   align-items: center;
   text-align: center;
   padding: 1rem;
+}
+.h2-title > h2 {
+  color: var(--color-neutral);
+  font-family: var(--font-header-2-3);
+  font-size: 5rem;
+  font-weight: 800;
+  text-shadow: none;
 }
 main {
   display: flex;
@@ -108,6 +105,17 @@ main {
   text-align: center;
   padding: 1rem;
 }
+.links {
+  border-radius: 1rem;
+  box-shadow: var(--shadow);
+  cursor: pointer;
+  font-family: "IBM Plex Mono", sans-serif;
+  font-size: 1rem;
+  font-weight: 600;
+  padding: 0.5rem 1rem;
+  text-align: center;
+}
+
 footer {
   display: flex;
   justify-content: start;
@@ -119,6 +127,7 @@ footer {
   height: auto;
   background-color: var(--color-secondary);
 }
+
 @media (min-width: 16rem) {
   body {
     font-size: 1rem;
